+++ Backend/platform/src/main/java/com/mathsmastery/platform/repository/UnitRepository.java
@@ -3,7 +3,7 @@ package com.mathsmastery.platform.repository;
 import com.mathsmastery.platform.model.Unit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnitRepository extends JpaRepository<Unit, Long> {
+public interface UnitRepository extends JpaRepository<Unit, Integer> {
 
 
 }
